backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.20.0
pydantic>=2.0.0
pypdf>=3.17.0
tiktoken>=0.7.0
faiss-cpu>=1.7.4
numpy>=1.24.0
google-generativeai>=0.3.0
reportlab>=4.0.0
langchain-text-splitters
python-dotenv
